Moved views to separate folder
