+++ src/slsp/features/CTHandler.ts
@@ -3,9 +3,9 @@
 import * as vscode from 'vscode';
 import { commands, ExtensionContext, Uri, window, WorkspaceFolder } from "vscode";
 import { CancellationTokenSource, LSPErrorCodes, WorkDoneProgress } from "vscode-languageclient";
-import { CTTestCase, CTSymbol, CTFilterOption, CTTracesParams, CTTracesRequest, CTGenerateParams, CTGenerateRequest, CTExecuteParams, CTExecuteRequest, NumberRange } from "../../protocol/combinatorialTesting";
-import { SpecificationLanguageClient } from "../../../SpecificationLanguageClient";
-import { CTTreeView } from './CTTreeView';
+import { CTTestCase, CTSymbol, CTFilterOption, CTTracesParams, CTTracesRequest, CTGenerateParams, CTGenerateRequest, CTExecuteParams, CTExecuteRequest, NumberRange } from "../protocol/combinatorialTesting";
+import { SpecificationLanguageClient } from "../../SpecificationLanguageClient";
+import { CTTreeView } from '../views/CTTreeView';
 
 export class CTHandler {
     private _ctTreeView: CTTreeView;
