+++ src/slsp/views/CTDataProvider.ts
@@ -2,8 +2,8 @@
 
 import { Event, EventEmitter, ExtensionContext, TreeDataProvider, TreeItem, TreeItemCollapsibleState } from 'vscode';
 import { CTTreeView } from './CTTreeView';
-import { NumberRange, VerdictKind } from '../../protocol/combinatorialTesting';
-import { Icons } from '../../../Icons'
+import { NumberRange, VerdictKind } from '../protocol/combinatorialTesting';
+import { Icons } from '../../Icons'
 
 export class CTTreeDataProvider implements TreeDataProvider<TestViewElement> {
